requirements.txt
fastapi
uvicorn
pymongo
motor
python-jose
passlib[bcrypt]
jinja2
python-multipart
pydantic<2.0
